requirements.txt
# Django Core
Django==5.0.1
django-environ==0.11.2
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
django-extensions==3.2.3

# Authentication & Security
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Vector Database Support
pgvector==0.2.4

# Async Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0

# Web Scraping and Crawling
crawl4ai>=0.6.0
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
aiohttp==3.9.1
playwright>=1.40.0

# AI/ML Libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
openai==1.3.7
sentence-transformers==2.2.2
langchain==0.0.340
numpy==1.24.3
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0

# Development
django-debug-toolbar==4.2.0
pytest-django==4.7.0
factory-boy==3.3.0 